feat(mockfile): create a mock chart for jll project
