+++ package.json
@@ -11,6 +11,7 @@
     "@types/react": "^17.0.3",
     "@types/react-dom": "^17.0.3",
     "@types/react-highcharts": "^16.0.2",
+    "highcharts": "^9.0.1",
     "highcharts-react-official": "^3.0.0",
     "highcharts-regression": "^2.2.0",
     "react": "^17.0.2",
